Fix bugs: Empty Event, return type for unedit Event, update blood quantity for approve request

+++ backend/Controllers/RequestController.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using backend.Models;
 using backend.Repositories;
@@ -16,13 +15,15 @@ namespace backend.Controllers
     {
         private readonly IRequestRepository _requestRepository;
         private readonly IHospitalRepository _hospitalRepository;
+        private readonly IBloodRepository _bloodRepository;
 
-        public RequestController(IRequestRepository requestRepository, IHospitalRepository hospitalRepository)
+        public RequestController(IRequestRepository requestRepository, IHospitalRepository hospitalRepository,
+            IBloodRepository bloodRepository)
         {
             _requestRepository = requestRepository;
             _hospitalRepository = hospitalRepository;
+            _bloodRepository = bloodRepository;
         }
-        
 
         [HttpPut("approve")]
         public async Task<IActionResult> ApproveRequest(IEnumerable<Request> listRequest)
@@ -31,17 +32,29 @@ namespace backend.Controllers
             {
                 foreach (var request in listRequest)
                 {
-                    var result = await _requestRepository.Get(request._id);
-                    result.ApproveStatus = 1;
-                    await _requestRepository.Update(request._id, result);
-                }
+                    // Check empty or approved request.
+                    var existRequest = await _requestRepository.Get(request._id);
+                    if (existRequest == null) throw new Exception();
+                    if (existRequest.Status == 1) throw new Exception("approved");
 
+                    // Check valid quantity request.
+                    var blood = _bloodRepository.GetByNameAndType(existRequest.Blood.Name, existRequest.Blood.Type);
+                    if (existRequest.Quantity > blood.Result.quantity) throw new Exception("quantity");
+
+                    await _bloodRepository.UpdateQuantity(existRequest.Blood.Name, existRequest.Blood.Type,
+                        -existRequest.Quantity);
+                    _requestRepository.ApproveRequest(request);
+                }
                 return Ok("Approve request successfully!");
             }
             catch (Exception e)
             {
-                Console.WriteLine(e);
-                return BadRequest("Request ID error!");
+                return e.Message switch
+                {
+                    "quantity" => BadRequest("Not enough blood!"),
+                    "approved" => BadRequest("Request already approved!"),
+                    _ => BadRequest("Request ID error!")
+                };
             }
         }
 
@@ -52,18 +65,21 @@ namespace backend.Controllers
             {
                 foreach (var request in listRequest)
                 {
-                    var result = await _requestRepository.Get(request._id);
-                    result.ApproveStatus = -1;
-                    result.RejectReason = request.RejectReason;
-                    await _requestRepository.Update(result._id, result);
-                }
+                    var existRequest = await _requestRepository.Get(request._id);
+                    if (existRequest == null) throw new Exception();
+                    if (existRequest.Status == -1) throw new Exception("rejected");
 
+                    _requestRepository.RejectRequest(request);
+                }
                 return Ok("Reject request successfully!");
             }
             catch (Exception e)
             {
-                Console.WriteLine(e);
-                return BadRequest("Request ID error!");
+                return e.Message switch
+                {
+                    "rejected" => BadRequest("Request already rejected!"),
+                    _ => BadRequest("Request ID error!")
+                };
             }
         }
 
@@ -90,6 +106,11 @@ namespace backend.Controllers
             try
             {
                 var result = await _requestRepository.Get(id);
+                if (result == null)
+                {
+                    throw new Exception();
+                }
+
                 return Ok(result);
             }
             catch (Exception e)
@@ -105,6 +126,32 @@ namespace backend.Controllers
             try
             {
                 var result = await _requestRepository.Get();
+                if (result == null)
+                {
+                    throw new Exception();
+                }
+
+                return Ok(result);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                return BadRequest("Get request error!");
+            }
+        }
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(string id, Request request)
+        {
+            try
+            {
+                var exist = await _requestRepository.Get();
+                if (exist == null)
+                {
+                    throw new Exception();
+                }
+
+                var result = await _requestRepository.Update(id, request);
                 return Ok(result);
             }
             catch (Exception e)
@@ -114,33 +161,24 @@ namespace backend.Controllers
             }
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> Update(string id, Request request)
-        {
-            try
-            {
-                var result = await _requestRepository.Update(id, request);
-                return Ok(result);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e);
-                return BadRequest("Hospital ID error!");
-            }
-        }
-
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(string id)
         {
             try
             {
+                var exist = await _requestRepository.Get();
+                if (exist == null)
+                {
+                    throw new Exception();
+                }
+
                 var result = await _requestRepository.Delete(id);
                 return Ok(result);
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
-                return BadRequest("Hospital ID error!");
+                return BadRequest("Request ID error!");
             }
         }
     }
